main loop and directories
